+++ test/HatTrick.DbEx.MsSql.Test.Database/Executor/Arithmetic.cs
@@ -12,7 +12,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
     {
         [Theory]
         [MsSqlVersions.AllVersions]
-        public void Does_arithmetic_order_of_precedence_succeed(int version, decimal expected = 36m)
+        public void Does_arithmetic_order_of_precedence_succeed(int version, double expected = 36)
         {
             //given
             ConfigureForMsSqlVersion(version);
@@ -20,10 +20,10 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
             var exp = db.SelectOne(
                     (dbo.Purchase.TotalPurchaseAmount + 2) * 3
                 ).From(dbo.Purchase)
-                .Where(dbo.Purchase.TotalPurchaseAmount == 10m);
+                .Where(dbo.Purchase.TotalPurchaseAmount == 10);
 
             //when               
-            decimal result = exp.Execute();
+            double result = exp.Execute();
 
             //then
             result.Should().Be(expected);
@@ -31,7 +31,7 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
 
         [Theory]
         [MsSqlVersions.AllVersions]
-        public void Does_arithmetic_order_of_precedence_reversed_succeed(int version, decimal expected = 36m)
+        public void Does_arithmetic_order_of_precedence_reversed_succeed(int version, double expected = 36)
         {
             //given
             ConfigureForMsSqlVersion(version);
@@ -39,10 +39,10 @@ namespace HatTrick.DbEx.MsSql.Test.Database.Executor
             var exp = db.SelectOne(
                     3 * (dbo.Purchase.TotalPurchaseAmount + 2)
                 ).From(dbo.Purchase)
-                .Where(dbo.Purchase.TotalPurchaseAmount == 10m);
+                .Where(dbo.Purchase.TotalPurchaseAmount == 10);
 
             //when               
-            decimal result = exp.Execute();
+            double result = exp.Execute();
 
             //then
             result.Should().Be(expected);
